+++ tecMedical/src/main/java/com/cib/tecMedical/repository/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cib.tecMedical.entidades.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Long>{
+public interface ClienteRepository extends JpaRepository<Cliente, Integer>{
 
 }
